pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langbridge"
version = "0.1.0"
description = "Langbridge orchestration platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi[standard]==0.117.1",
  "uvicorn[standard]==0.32.0",
  "opentelemetry-api==1.39.1",
  "opentelemetry-sdk==1.39.1",
  "opentelemetry-exporter-otlp==1.39.1",
  "opentelemetry-instrumentation-fastapi==0.60b1",
  "opentelemetry-instrumentation-logging==0.60b1",
  "dependency-injector==4.48.2",
  "sqlalchemy==2.0.43",
  "aiosqlite==0.20.0",
  "asyncpg==0.30.0",
  "pydantic-settings==2.11.0",
  "python-jose[cryptography]==3.3.0",
  "httpx==0.27.2",
  "PyYAML==6.0.2",
  "authlib==1.3.1",
  "psycopg[binary]==3.2.3",
  "itsdangerous==2.2.0",
  "snowflake-connector-python==3.18.0",
  "cryptography==46.0.0",
  "sqlglot==27.25.2",
  "openai==2.2.0",
  "anthropic>=0.39.0",
  "langchain-openai>=0.1.0",
  "langchain-anthropic>=0.1.0",
  "langchain==0.3.27",
  "mcp==1.17.0",
  "pandas==2.3.3",
  "psycopg2==2.9.11",
  "numpy==1.26.4",
  "faiss-cpu==1.8.0",
  "trino==0.336.0",
  "qdrant-client==1.9.1",
]

[tool.setuptools.packages.find]
where = ["langbridge"]
exclude = ["venv*", "__pycache__*", "tests*"]
